feat: Implement live scoreboard functionality; add API and hooks for fetching live competition data
