+++ constants/url/url.ts
@@ -29,3 +29,6 @@ export const GET_MY_EVENT_REGISTRATION_FORM_URL =
 export const INITIALIZE_ATTEMPTS = "/attempt/init-lifts";
 // constants/url/liftAttemptUrls.ts
 export const GET_LIFT_ATTEMPTS_URL = "/attempt";
+
+//
+export const GET_LIVE_SCOREBOARD_URL = "/live-scoreboard/:eventId";
